facets: Extract `sortBy` into hook
No change in behavior, just pulling the logic that determines the `sortBy`
function into a hook. We'll add some more logic to this hook in the next few
commits.
